More consistent handling of internal types, minimizes unnecessary rounding.
Hue, saturation, lightness and alpha are all stored as floats, retrieved as floats.
Hue saturation and lightness are rounded off at 5 digits.
Red, green and blue are rounded and stored as ints in constructor, retrieved as ints.
Removes a number of type conversions (intval and floatval calls).
Removes a number of return types on map functions.
Updates a number of tests where we were previously off by one from SASS.

+++ src/Color.php
@@ -76,12 +76,12 @@ class Color implements ColorInterface
         return $this->hsl;
     }
 
-    public function getHue() : int
+    public function getHue() : float
     {
         return $this->hsl->getHue();
     }
 
-    public function getLightness() : int
+    public function getLightness() : float
     {
         return $this->hsl->getLightness();
     }
@@ -91,6 +91,9 @@ class Color implements ColorInterface
         return $this->rgb->getName();
     }
 
+    /**
+     * @todo Should this method be on the Rgb class?
+     */
     public function getPerceivedBrightness() : int
     {
         $weightedRed = 0.299 * pow($this->getRed() / 255, 2);
@@ -110,7 +113,7 @@ class Color implements ColorInterface
         return $this->rgb;
     }
 
-    public function getSaturation() : int
+    public function getSaturation() : float
     {
         return $this->hsl->getSaturation();
     }
